# src/main/java/com/crt/server/dto/UserDTO.java
package com.crt.server.dto;

import com.crt.server.model.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserDTO {
    private UUID id;
    private String name;
    private String email;
    private String phone;
    private String username;
    private Role role;
    private Boolean isFirstLogin;
    private Boolean isActive;}